requirements.txt
numpy>=1.24.0
pandas>=2.0.0
scikit-learn>=1.3.0
matplotlib>=3.7.0
seaborn>=0.12.0
mlflow>=2.10.0
pyarrow>=14.0.0
pyyaml>=6.0.1
kagglehub>=0.2.0
jupyterlab>=4.0.0
